requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# HTTP and requests
requests>=2.31.0
httpx>=0.25.0

# YAML parsing
PyYAML>=6.0.1

# CLI and utilities
click>=8.1.7

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
moto>=4.2.0  # AWS mocking for tests
